requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
twilio==8.5.0
uritemplate==4.1.1
urllib3==2.0.4
yarl==1.9.2
pytest-django
aiogram
pytest-cov
bs4
lxml